Reorder Resume Uploading And Downloading To Be In Resources Controller
