+++ nextstep-backend/src/routes/resume_routes.ts
@@ -5,16 +5,8 @@ import { authenticateToken } from "../middleware/auth";
 
 const router = express.Router();
 
-// Upload and score a resume
-router.post('/upload', authenticateToken, (req: Request, res: Response) => 
-    ResumeController.uploadAndScoreResume(req as CustomRequest, res));
-
 // Get score for an existing resume
 router.get('/score/:filename', authenticateToken, (req: Request, res: Response) => 
     ResumeController.getResumeScore(req, res));
-
-// Get resume file
-router.get('/file/:filename', authenticateToken, (req: Request, res: Response) => 
-    ResumeController.getResumeFile(req, res));
 
 export default router;
